% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alt}
\alias{alt}
\title{Alanine Aminotransferase (ALT) values.}
\format{A data frame with 240 rows and 2 variables:
\describe{
  \item{value}{ALT values (U/L)}
  \item{sex}{student sex}
  ...
}}
\usage{
alt
}
\description{
A dataset containing the ALT values of 240 medical students
(120 males and 120 females) of the University of Virginia
(1987 and 1988).
}
\references{
CLSI (...)
}
\keyword{datasets}

