% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefInterval.R
\name{AnalyteDistPlot}
\alias{AnalyteDistPlot}
\title{Analyte distribution plot}
\usage{
AnalyteDistPlot(x, analyte.name = "Analyte", main = "",
  probability = TRUE)
}
\arguments{
\item{x}{a numeric vector with the analyte values.}

\item{analyte.name}{an optional character string giving the analyte name.}

\item{main}{an overall title for the plot.}

\item{logical;}{probability if TRUE, probability densities are plotted
and a normal density curve is added computed with sample
\code{\link[base]{mean}} and standard deviation (\code{\link[base]{sd}});
if FALSE, the histogram graphic is a representation of frequencies and
no curve is plotted.}
}
\value{
A histogram plot of the analyte.
}
\description{
Produces a histogram plot of the analyte.
}
\examples{
data(ca)
AnalyteDistPlot(ca$value)
AnalyteDistPlot(ca$value, analyte.name = "Calcium (mg/dL)")

data(alt)
AnalyteDistPlot(alt$value[alt$sex == "m"],
analyte.name = "Alanine Aminotrasferase (U/L)", main = "Male", probability = FALSE)
}

