% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefInterval.R
\name{AnalyteSummary}
\alias{AnalyteSummary}
\title{Analyte summary statistics}
\usage{
AnalyteSummary(x)
}
\arguments{
\item{x}{a numeric vector with the analyte values.}
}
\value{
A vector with the analyte summary statistics: valid sample size, mean, standard deviation
range, median, first and third quartiles, and number of missing observations.
}
\description{
Describes analyte distribution with summary statistics.
}
\examples{
data(ca)
AnalyteSummary(ca$value)

data(alt)
AnalyteSummary(alt$value[alt$sex == "f"])
}

