% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ca}
\alias{ca}
\title{Calcium values.}
\format{A data frame with 240 rows and 2 variables:
\describe{
  \item{value}{calcium values (mg/dL)}
  \item{sex}{student sex}
  ...
}}
\usage{
ca
}
\description{
A dataset containing the calcium values of 240 medical students
(120 males and 120 females) of the University of Virginia
(1987 and 1988).
}
\references{
CLSI (...)
}
\keyword{datasets}

