% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{casub}
\alias{casub}
\title{Calcium values subset.}
\format{A numeric vector of length 20.}
\usage{
casub
}
\description{
A dataset containing a subset of 20 calcium values of the medical students
of the University of Virginia (1987 and 1988).
}
\references{
CLSI (...)
}
\keyword{datasets}

