% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefInterval.R
\name{RefInterval}
\alias{RefInterval}
\title{Analyte reference intervals}
\usage{
RefInterval(data = NULL, x, type.ri = "all", conf.ci = 0.9, B = 1000,
  analyte.name = "Analyte", partition = NULL)
}
\arguments{
\item{data}{a dataframe.}

\item{x}{a numeric vector with the analyte values.}

\item{type.ri}{a character string giving a type for computing the reference interval.
This must be one of the strings "all" (default), "non-parametric", "parametric", "bootstrap",
"robust".}

\item{conf.ci}{numeric; the confidence level for computing confidence intervals for the
limits of the reference interval. This must be between zero and one.}

\item{B}{the number of bootstrap replicates for computing confidence intervals for the limits
of reference interval, and reference interval type "bootstrap".}

\item{analyte.name}{an optional character string giving the analyte name.}

\item{partition}{an optional character string giving the partition variable name. Default is
NULL. If partition is not NULL, then reference intervals for each level of the partition
are computed.}
}
\value{
Summary stistics, distribution plots, and reference intervals with confidence intervals.
}
\description{
Produces reference intervals for the analyte.
}
\examples{
data(ca)
RefInterval(data = ca, x = "value", analyte.name = "Calcium (mg/dL)")
RefInterval(data = ca, x = "value",
type.ri = "non-parametric", analyte.name = "Calcium (mg/dL)",
partition = "sex")

# Robust ref. intervals for small data sets
data(casub)
RefInterval(data = casub, x = "value",
type.ri = "robust", analyte.name = "Calcium (mg/dL)")
}
\references{
CLSI (...)
}

